+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed `conversion` module was private (#42)
 * Fixed `Hex::line_to` returning `(0, 0)` when both ends are identical (#43)
+* Deprecated `Hex::to_array3` in favor of `Hex::to_cubic_array`
 
 ## 0.4.1